Flatten paste and paste-impl into one crate
The separation between implementation crate and declaration crate is no
longer needed after dropping proc-macro-hack.
