--- impl/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "paste-impl"
-version = "0.1.18"
-authors = ["David Tolnay <dtolnay@gmail.com>"]
-edition = "2018"
-license = "MIT OR Apache-2.0"
-description = "Implementation detail of the `paste` crate"
-repository = "https://github.com/dtolnay/paste"
-
-[lib]
-proc-macro = true
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
