--- impl/LICENSE-MIT
@@ -1,1 +1,0 @@
-../LICENSE-MIT
